refactor: split ChatTopBar into directory with extracted badge components
Moved ChatTopBar.tsx into components/ChatTopBar/ directory and
extracted WorkspaceBadge and PullRequestBadge into their own files.
Also fixed a stray {" "} in the JSX.

The PR badge logic (url parsing, hasPR check, render-or-null) now
lives in PullRequestBadge, which simplifies ChatTopBar and makes
the badge independently testable.
