+++ site/src/pages/AgentsPage/components/ChatTopBar/ChatTopBar.stories.tsx
